+++ src/user/uploads.js
@@ -12,7 +12,7 @@ const batch = require('../batch');
 
 const md5 = filename => crypto.createHash('md5').update(filename).digest('hex');
 const _getFullPath = relativePath => path.resolve(nconf.get('upload_path'), relativePath);
-const _validatePath = async (relativePaths) => {
+const _validatePath = async relativePaths => {
     if (typeof relativePaths === 'string') {
         relativePaths = [relativePaths];
     } else if (!Array.isArray(relativePaths)) {
@@ -30,10 +30,7 @@ const _validatePath = async (relativePaths) => {
 module.exports = function (User) {
     User.associateUpload = async (uid, relativePath) => {
         await _validatePath(relativePath);
-        await Promise.all([
-            db.sortedSetAdd(`uid:${uid}:uploads`, Date.now(), relativePath),
-            db.setObjectField(`upload:${md5(relativePath)}`, 'uid', uid),
-        ]);
+        await Promise.all([db.sortedSetAdd(`uid:${uid}:uploads`, Date.now(), relativePath), db.setObjectField(`upload:${md5(relativePath)}`, 'uid', uid)]);
     };
 
     User.deleteUpload = async function (callerUid, uid, uploadNames) {
@@ -45,46 +42,45 @@ module.exports = function (User) {
 
         await _validatePath(uploadNames);
 
-        const [isUsersUpload, isAdminOrGlobalMod] = await Promise.all([
-            db.isSortedSetMembers(`uid:${callerUid}:uploads`, uploadNames),
-            User.isAdminOrGlobalMod(callerUid),
-        ]);
+        const [isUsersUpload, isAdminOrGlobalMod] = await Promise.all([db.isSortedSetMembers(`uid:${callerUid}:uploads`, uploadNames), User.isAdminOrGlobalMod(callerUid)]);
         if (!isAdminOrGlobalMod && !isUsersUpload.every(Boolean)) {
             throw new Error('[[error:no-privileges]]');
         }
 
-        await batch.processArray(uploadNames, async (uploadNames) => {
-            const fullPaths = uploadNames.map(path => _getFullPath(path));
+        await batch.processArray(
+            uploadNames,
+            async uploadNames => {
+                const fullPaths = uploadNames.map(path => _getFullPath(path));
 
-            await Promise.all(fullPaths.map(async (fullPath, idx) => {
-                winston.verbose(`[user/deleteUpload] Deleting ${uploadNames[idx]}`);
-                await Promise.all([
-                    file.delete(fullPath),
-                    file.delete(file.appendToFileName(fullPath, '-resized')),
-                ]);
-                await Promise.all([
-                    db.sortedSetRemove(`uid:${uid}:uploads`, uploadNames[idx]),
-                    db.delete(`upload:${md5(uploadNames[idx])}`),
-                ]);
-            }));
+                await Promise.all(
+                    fullPaths.map(async (fullPath, idx) => {
+                        winston.verbose(`[user/deleteUpload] Deleting ${uploadNames[idx]}`);
+                        await Promise.all([file.delete(fullPath), file.delete(file.appendToFileName(fullPath, '-resized'))]);
+                        await Promise.all([db.sortedSetRemove(`uid:${uid}:uploads`, uploadNames[idx]), db.delete(`upload:${md5(uploadNames[idx])}`)]);
+                    })
+                );
 
-            // Dissociate the upload from pids, if any
-            const pids = await db.getSortedSetsMembers(uploadNames.map(relativePath => `upload:${md5(relativePath)}:pids`));
-            await Promise.all(pids.map(async (pids, idx) => Promise.all(
-                pids.map(async pid => posts.uploads.dissociate(pid, uploadNames[idx]))
-            )));
-        }, { batch: 50 });
+                // Dissociate the upload from pids, if any
+                const pids = await db.getSortedSetsMembers(uploadNames.map(relativePath => `upload:${md5(relativePath)}:pids`));
+                await Promise.all(pids.map(async (pids, idx) => Promise.all(pids.map(async pid => posts.uploads.dissociate(pid, uploadNames[idx])))));
+            },
+            { batch: 50 }
+        );
     };
 
     User.collateUploads = async function (uid, archive) {
-        await batch.processSortedSet(`uid:${uid}:uploads`, (files, next) => {
-            files.forEach((file) => {
-                archive.file(_getFullPath(file), {
-                    name: path.basename(file),
+        await batch.processSortedSet(
+            `uid:${uid}:uploads`,
+            (files, next) => {
+                files.forEach(file => {
+                    archive.file(_getFullPath(file), {
+                        name: path.basename(file),
+                    });
                 });
-            });
 
-            setImmediate(next);
-        }, { batch: 100 });
+                setImmediate(next);
+            },
+            { batch: 100 }
+        );
     };
 };
